Update UI to 2.8+
In 2.8+ the UI is single column and right aligned. This patch updates the freesound ui accordingly. 
There was a small change from user_preferences to preferences to get it to work. 

Hmm... I can't paste an image in here of the ui. 

The duration to/frem seems broken - if they can be fixed the value of the current selected file could maybe be added automatically upon selection?

NB. Blender can't install add-ons with '-' in the file name so currently can't the 2.0 version be installed as zip.